daemons-lircd: Set the protocol for rc device used.
There are tons of descriptions how to echo lirc >
/sys/class/rc/rcX/protocols to make lirc work as intended. This is
needlessly complicated. This patch tries to locate the device used
presuming that if the device /dev/lircN actually refers to
/sys/class/rc/rcX there is a link /sys/class/rc/rcX/lircN. This is
not part of the stable kernel interface as I understand, but should
be good enough for now. If found, 'lirc' is just written to the
corresponding 'protocols' file.

If the driver is devinput, no protocol change is done. If the device
isn't an rc device the logic will try to set protocol but silently
fails since it will not find  a device link in the rc directories.

Existing wrapper scripts should not be a problem, doing this twice
does not make any difference.

The systemd installation is of course so much simpler not having
to cope with this stuff.

Conflicts:
	daemons/lircd.c
	systemd/lircd.service