doc: Revamp build system to comply with autotool defaults.
The doc was strange with the build done in special targets,
not invoked on ordinary "make all". As a  consequence, simple
"make distcheck" fails in a number of more or less interesting
ways.

New Makefile declares the docs as standard DATA variables handled
by automake and built by 'make all'.  This one actually passes
"make distcheck". The downside is that Makefile.am is full of
long lists.

New version does not distribute the manpages. In general, this is a
bad idea since the final layout might depend on configure. In this
case, some pages are not even installed depending on configuration.

Adds a new build dependency on help2man dince the docs re always built.