config: Adding modprobe support.
When configuring lircd, it is configured with a number of kernel
modules (possibly 0). There is up to this point no support for
this. This commit adds a modules= entry in lirc_options.conf
and a script lircd-modprobe which is intended to run before
lircd e.g., in a systemd PreExec tag. Since it reads the
configuration file it's safe to run even if there is no
modules to load.

Updating lirc_options from configuration scripts should be
straightforward. Adding modprobe options to /etc/modprobe.d is
also possible, but with lot's of caveats when installed configs
collides with existing ones.