plugins: mplay bugfixes and overhaul.
The mplay2 plugin for the Moneual MonCaso 312/320 pc cases does not work
correctly. Some keys (for example REC, code 0x11) are, as far as I
understand it, recognized as control characters and then ignored. Also
the volume knob is not supported.

There is a pretty old thread at ubuntuforums.org
(http://ubuntuforums.org/showthread.php?t=1536934&p=12120799#post12120799)
where the user Womiha provides a fix (which I tested and works), but it
was never added to lirc. This patch is basically the same, rebased to
0.9.2.